Tvws_app/requirements.txt
mysql==0.0.3
mysqlclient==2.1.1

dj-database-url==1.0.0

django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-phonenumber-field==6.3.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0

grpcio==1.40.0
gunicorn==20.1.0
heroku==0.1.4


mysqlclient
nest-asyncio

xlwt==1.3.0
django
psycopg2-binary
djangorestframework
django-filter
dj-rest-auth
django-cors-headers
drf-yasg
django-private-storage